requirements.txt
torch>=2.0.0
transformers>=4.37.0
peft>=0.8.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0

pillow>=9.5.0

numpy>=1.24.0
tqdm>=4.65.0
datasets>=2.14.0

wandb>=0.15.0
tensorboard>=2.14.0

jupyter>=1.0.0
ipywidgets>=8.0.0
qwen-vl-utils==0.0.8

gigachat
python-dotenv
langchain_core
langchain_gigachat

google-images-download
selenium 
webdriver-manager